// Workshop/frontend/src/app/models/users.ts
export class User{
    ime: string;
    prezime: string;
    lozinka: string;
    korisnickoIme: string;
    email: string;
    telefon: string;
    nazivOrganizacije: string;
    drzava: string;
    grad: string;
    postanskiBroj: string;
    ulica: string;
    broj: number;
    maticniBrojOrganizacije: string;
    tip: number;
    slika: string;
    status:string;

}